requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alpha-vantage==2.3.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
docutils==0.21.2
filelock==3.14.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.5.0
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
huggingface-hub==0.23.0
idna==3.7
isodate==0.6.1
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
rsa==4.9
safetensors==0.4.3
six==1.16.0
soupsieve==2.5
statistics==1.0.3.5
sympy==1.12
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
transformers==4.41.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vaderSentiment==3.3.2
yarl==1.9.4
youtube-transcript-api==0.6.2
